Add filter for bogus serial ports, and centralize that determination.
In my list of ports is a bunch of distractors (like /dev/ttyS0) in
addition to /dev/ttyUSB0, and the order always changes. I noticed
that /dev/ttyUSB0 was the only one with a "properties" object,
which includes the vid (Vendor ID) of the USB device. So let's
filter for that special record.

Because the filter is now a bit elaborate, we now calculate it in
the Makefile and then pass it to the webapp and arduino-cli as a
variable. Port detection is now removed from motorctl.py.